Move Classpath to util module
util will be a shared dependency of runtime / interp on the one hand,
and the upcoming compiler module on the other. Moving Classpath here
allows it to be used by both.
