+++ amm/interp/src/main/scala-2.12/ammonite/interp/CompilerCompatibility.scala
@@ -1,6 +1,6 @@
 package ammonite.interp
 
-import ammonite.runtime.Classpath
+import ammonite.util.Classpath
 
 import scala.reflect.internal.util.Position
 import scala.reflect.io.FileZipArchive
