+++ amm/interp/src/main/scala/ammonite/interp/Compiler.scala
@@ -3,8 +3,7 @@ package ammonite.interp
 
 import java.io.OutputStream
 
-import ammonite.runtime.Classpath
-import ammonite.util.{ImportData, Imports, Printer}
+import ammonite.util.{Classpath, ImportData, Imports, Printer}
 import ammonite.util.Util.newLine
 
 import scala.collection.mutable
