+++ amm/util/src/main/scala/ammonite/util/Classpath.scala
@@ -1,4 +1,4 @@
-package ammonite.runtime
+package ammonite.util
 
 import java.io.File
 import java.net.URL
